Renamed baw service files
- Added `.service` to api interceptor
- Renamed BawApiService file
