+++ src/app/component/home/home.component.spec.ts
@@ -15,8 +15,8 @@ import { delay } from "rxjs/operators";
 import { testBawServices } from "src/app/app.helper";
 import { Project } from "src/app/models/Project";
 import { AppConfigService } from "src/app/services/app-config/app-config.service";
-import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor";
-import { Filters } from "src/app/services/baw-api/base-api.service";
+import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor.service";
+import { Filters } from "src/app/services/baw-api/baw-api.service";
 import { ProjectsService } from "src/app/services/baw-api/projects.service";
 import { SecurityService } from "src/app/services/baw-api/security.service";
 import { SharedModule } from "../shared/shared.module";
