+++ src/app/component/profile/pages/profile/their-profile.component.ts
@@ -10,7 +10,7 @@ import { ImageSizes } from "src/app/interfaces/apiInterfaces";
 import { AnyMenuItem, MenuLink } from "src/app/interfaces/menusInterfaces";
 import { User } from "src/app/models/User";
 import { AccountService } from "src/app/services/baw-api/account.service";
-import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor";
+import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor.service";
 import {
   theirEditProfileMenuItem,
   theirProfileCategory,
