+++ src/app/component/projects/pages/list/list.component.spec.ts
@@ -11,7 +11,7 @@ import { Subject } from "rxjs";
 import { testBawServices } from "src/app/app.helper";
 import { SharedModule } from "src/app/component/shared/shared.module";
 import { Project } from "src/app/models/Project";
-import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor";
+import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor.service";
 import { ProjectsService } from "src/app/services/baw-api/projects.service";
 import { ListComponent } from "./list.component";
