+++ src/app/component/security/pages/register/register.component.ts
@@ -2,8 +2,7 @@ import { Component, OnDestroy, OnInit } from "@angular/core";
 import { Subject } from "rxjs";
 import { PageComponent } from "src/app/helpers/page/pageComponent";
 import { Page } from "src/app/helpers/page/pageDecorator";
-import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor";
-import { BawApiService } from "src/app/services/baw-api/base-api.service";
+import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor.service";
 import { SecurityService } from "src/app/services/baw-api/security.service";
 import { registerMenuItem, securityCategory } from "../../security.menus";
 import data from "./register.json";
