+++ src/app/component/shared/action-menu/action-menu.component.ts
@@ -6,7 +6,7 @@ import { takeUntil } from "rxjs/operators";
 import { DefaultMenu } from "src/app/helpers/page/defaultMenus";
 import { PageInfo } from "src/app/helpers/page/pageInfo";
 import { AnyMenuItem, LabelAndIcon } from "src/app/interfaces/menusInterfaces";
-import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor";
+import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor.service";
 import { WidgetMenuItem } from "../widget/widgetItem";
 
 @Component({
