+++ src/app/component/shared/error-handler/error-handler.component.ts
@@ -5,8 +5,8 @@ import {
   OnChanges,
   OnInit
 } from "@angular/core";
-import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor";
-import { apiReturnCodes } from "src/app/services/baw-api/base-api.service";
+import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor.service";
+import { apiReturnCodes } from "src/app/services/baw-api/baw-api.service";
 
 @Component({
   selector: "app-error-handler",
