+++ src/app/component/shared/menu/menu.component.ts
@@ -19,7 +19,6 @@ import {
   LabelAndIcon
 } from "src/app/interfaces/menusInterfaces";
 import { SessionUser } from "src/app/models/User";
-import { BawApiService } from "src/app/services/baw-api/base-api.service";
 import { SecurityService } from "src/app/services/baw-api/security.service";
 import { WidgetComponent } from "../widget/widget.component";
 import { WidgetDirective } from "../widget/widget.directive";
