+++ src/app/services/baw-api/user.service.spec.ts
@@ -7,14 +7,14 @@ import { RouterTestingModule } from "@angular/router/testing";
 import { Subject } from "rxjs";
 import { testAppInitializer } from "src/app/app.helper";
 import { User } from "src/app/models/User";
-import { ApiErrorDetails } from "./api.interceptor";
-import { BawApiService } from "./base-api.service";
+import { ApiErrorDetails } from "./api.interceptor.service";
+import { BawApiService } from "./baw-api.service";
 import {
   apiErrorDetails,
   apiErrorInfoDetails,
   shouldNotFail,
   shouldNotSucceed
-} from "./base-api.service.spec";
+} from "./baw-api.service.spec";
 import { MockBawApiService } from "./mock/baseApiMock.service";
 import { UserService } from "./user.service";
